requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
PyMuPDF==1.24.13
Pillow==11.0.0
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
requests==2.32.3
